update dropdown button to include links to user token and readme
