+++ public/scss/globals.scss
@@ -25,14 +25,53 @@ body {
     margin-left: 25px;
   }
 
-  #logoutbutton {
-    position: absolute;
-    transform: translate(-45px, -10px);
+  .dropdownbutton {
+    position: relative;
     color: $primary-color;
     background-color: $background-color;
     border: none;
     border: none;
     outline: none;
+    width: 100%;
+    height: 35px;
+    flex: 1;
+    padding: 0px;
+    border-radius: 15px;
+    z-index: 9;
+  }
+
+  #logoutbutton {
+    border-bottom-right-radius: 0px;
+    border-bottom-left-radius: 0px;
+    border-bottom: 1px solid $secondary-color;
+  }
+
+  #getuuidbutton {
+    border-top-right-radius: 0px;
+    border-top-left-radius: 0px;
+    border-top: 1px solid $secondary-color;
+  }
+
+  #readmebutton {
+    border-top-right-radius: 0px;
+    border-top-left-radius: 0px;
+    border-bottom-right-radius: 0px;
+    border-bottom-left-radius: 0px;
+    border-bottom: 1px solid $secondary-color;
+    border-top: 1px solid $secondary-color;
+  }
+
+  #readmebutton:hover {
+    border-bottom: 1px solid $tertiary-color;
+    border-top: 1px solid $tertiary-color;
+  }
+
+  #logoutbutton:hover {
+    border-bottom: 1px solid $tertiary-color;
+  }
+
+  #getuuidbutton:hover {
+    border-top: 1px solid $tertiary-color;
   }
 
   #userbutton {
@@ -46,6 +85,7 @@ body {
     outline: none;
     z-index: 10;
     transition: 0.5s;
+    box-sizing: border-box;
   }
 
   #userbutton:hover {
@@ -55,12 +95,12 @@ body {
   .dropdown-menu {
     background-color: $background-color;
     border: 1px solid $secondary-color;
-    border-radius: 5px;
+    border-radius: 15px;
     position: absolute;
     top: 60px;
     right: 25px;
-    width: 85px;
-    height: 30px;
+    width: 125px;
+    height: 105px;
     opacity: 0;
     visibility: hidden;
     transform: translateY(-20px);
@@ -69,6 +109,27 @@ body {
     vertical-align: middle;
     align-items: center;
     transition: 0.5s ease;
+    z-index: 9;
+  }
+
+  .dropdown-menu ul {
+    position: relative;
+    height: 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    list-style: none;
+    padding-inline-start: 0px;
+    margin-block-start: 0em;
+    margin-block-end: 0em;
+    z-index: 9;
+  }
+
+  .menu ul li {
+    display: flex;
+    box-sizing: border-box;
+    border-radius: 15px;
+    z-index: 9;
   }
  
   .dropdown-menu.active {
@@ -76,14 +137,13 @@ body {
     visibility: visible;
     transform: translateY(0);
     top: 60px;
+    border: 1px solid $secondary-color;
+    border-radius: 15px;
+    z-index: 9;
   }
 
   .dropdown-menu:hover {
     border: 1px solid $tertiary-color;
-  }
-  
-  .dropdown-menu ul {
-    list-style: none;
   }
 
   .menu-trigger {
